Add files via upload
In this real-world setting, you will need to piece together a series of models to perform different tasks; for instance, the algorithm that detects humans in an image will be different from the CNN that infers dog breed.  There are many points of possible failure, and no perfect algorithm exists.  Your imperfect solution will nonetheless create a fun user experience!
 The Road Ahead
  We break the notebook into separate steps. 
  Import Datasets
  Detect Humans
  Detect Dogs
  Create a CNN to Classify Dog Breeds (from Scratch)
  Create a CNN to Classify Dog Breeds (using Transfer Learning)
  Write your Algorithm
  Test Your Algorithm